# Tier1 decoder speed optimizations